Change source RNG to 64bit for FillRandom

+++ Plugin/NativeVertexCacheTest/TestUtil.h
@@ -24,7 +24,7 @@ inline void FillRandom(T& vec, uint64_t state = 0x900dbeef, uint64_t inc = 0x876
     using ValueType = T::value_type;
     Pcg pcg { state, inc };
     for(auto& v : vec) {
-        v = static_cast<ValueType>(pcg.getUint32());
+        v = static_cast<ValueType>(pcg.getUint64());
     }
 }
